qemu: fix missing error reports in capabilities probing
The "virsh domcapabilities --arch ppc64" command will fail with no
error message set if qemu-system-ppc64 is not currently installed.

This is because virQEMUCapsCacheLookup() does not report any error
message if not capabilities can be obtained from the cache. Almost
all methods calling this expected an error to be set on failure.

Once that's fixed though, we see a further bug which is that
virQEMUCapsCacheLookupDefault() is passing a NULL binary path to
virQEMUCapsCacheLookup(), so we need to catch that too.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20200615135641.540502-1-berrange@redhat.com>
Reviewed-by: Daniel Henrique Barboza <danielhb413@gmail.com>