domain_conf: Move zPCI validation from formatter to validator
In 076591009ad a validation code was added to
virDomainDeviceInfoFormat() which reports an error if zPCI
address entered in was incomplete. But, there are two problems
with this approach.

The first problem is the placement of the code - it doesn't
belong into XML formatter rather than XML validator.

The second one is that the code doesn't check if the PCI address
it's validating has zPCI extension and thus the error is reported
for normal PCI addresses too, hence the change of the function
that's being called.

The second problem is addressed in d482cf6bef which doesn't
address the first problem.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <97d965a9842a6d90d01c94f2adae0df40fe05726.1593416559.git.mprivozn@redhat.com>