stream: Don't read block device twice
The aim of virFileInDataDetectZeroes() is to mimic virFileInData()
so that it can be used in sparse virStream with very little change
to logic. This was implemented in previous commits. These two
functions act alike - the passed FD is rewound back to the position
it was in when either of the functions was called. For regular
files (when virFileInData()) is used, this is not a problem -
rewinding an FD is usually cheap and data is read() only from data
section and it's read only once. But this is not the case for
virFileInDataDetectZeroes(), which doesn't use cheap lseek() to
learn data/hole sections, but uses read() and zero buffer
detection. Worst case scenario, virFileInDataDetectZeroes() reads
1MiB of bytes, finds it is non-zero data, rewinds FD that 1MiB back
only so that virFDStreamThreadDoRead() (in case of daemon) or
virshStreamSource() (in case of client) can read the same data
again.

Possibly, this is not that big problem because kernel is likely to
keep the data still in cache. But let's not rely on that and store
the buffer for later use if we learned it contains actual data.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <05ec2a612ad7f6a83b322fd81af491fae633b0db.1594150891.git.mprivozn@redhat.com>