util: Rework virFileFlock() to be unambiguous
The boolean parameters for lock/unlock and read/write (shared/exclusive) caused
a lot of confusion when reading the callers.  The new approach is explicit and
unambiguous.

While at it, also change the only caller so that it acquires an exclusive lock
as it should've been all the time.  This was caused because the function was
ambiguous since it was created in commit 5a0a5f7fb5f5, and due to that it was
misused in commit 657ddeff2313 and since then the lock being taken was shared
rather than exclusive.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <edb0e14faef54effca66178299b847f303c28361.1594387252.git.mkletzan@redhat.com>