meson: remove automake specific directives
EXTRA_DIST is not relevant because meson makes a git copy when creating
dist archive so everything tracked by git is part of dist tarball.

BUILD_SOURCES is not relevant as all dependencies should be explicitly
tracked.

The remaining ones are not converted to meson files as they are
automatically tracked by meson.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <4f0e4545b71d24e15f3538a1103a2327f5d025e8.1594891445.git.phrdina@redhat.com>